Fix a nasty deadlock between internal libproc process stops and Pwait().
At times (e.g. when a Ptrace() is issued), libproc needs to ensure that the
child process has stopped before returning.  This was interacting badly with the
multi-thread support code and the nonblocking Pwait()s issued by the
process-control DTrace thread in dt_proc_control().

The problem is that when the SIGSTOP is delivered, but before a Pwait() is
issued for it (which would be proxied through the main thread), the waitfd in
the main thread returns, noting that a process state change (to SIGSTOPped)
needs to be processed, but *not* that a proxied Pwait() is outstanding, because
it hasn't been issued yet.  We then block on the ptrace lock, and only proceed
once the proxied Pwait() is issued -- but that is too late, the poll() has
already returned, and the control thread consumes the SIGSTOP.  The control
thread then loops around, instantly returns from poll() with the proxy Pwait(),
initiates a blocking Pwait()... and waits forever, because we have already
consumed the SIGSTOP it was waiting for.  But we can't excise the blocking
Pwait()s from Ptrace() either, because then we might race the *other* way and
end up returning from Ptrace() not yet stopped, which would wreak all kinds of
havoc throughout the breakpoint code and rtld_db and could easily lead to the
child dying with a SIGTRAP as it slams into a breakpoint that hasn't been
properly overwritten with the original instruction yet.

We already have a 'pending stops' counter used to note the number of stop
requests that have been sent but not yet had a SIGSTOP received, but this is not
quite sufficient, because after the control thread has consumed the SIGSTOP this
counter falls to zero again, so Pwait() cannot tell that the subsequent blocking
Pwait() has no chance of success.  The solution is to add a new
awaiting_pending_stops flag, which is set whenever libproc sends a SIGSTOP and
unset when the SIGSTOP has been waited for.

Now we can treat this like a pthreads mutex, and rather than simply do a
blocking wait in Ptrace(), repatedly do a blocking wait until pending_stops is
unset (indicating that the stop was consumed, whether by our blocking wait or by
someone else calling Pwait(), we don't care which), or that the process has died
or been trace-stopped by someone else (again, as long as it's not running, it's
fine).  We also adjust Pwait_internal() to convert a blocking wait while no
pending stops are outstanding but we are awaiting at least one into a
nonblocking wait, so that if the SIGSTOP is consumed by another thread right
after we have checked the loop condition in Ptrace(), we don't end up blocking.

Finally, as a bit of extra paranoia, if the process is already known to not be
running for any other reason, or is known dead, there's no point doing a
blocking wait for it, since it will never return.  So reset the Pwait_internal()
blocking flag in this case too.

+++ libproc/Pcontrol.c
@@ -496,6 +496,23 @@ Pwait_internal(struct ps_prochandle *P, boolean_t block)
 	if (P->bkpt_halted)
 		return 0;
 
+	/*
+	 * If we are waiting for a pending stop, but pending_stops is zero, we
+	 * know an earlier Pwait() (whether from a breakpoint handler or from
+	 * another thread) has raced with this one and consumed the stop.  A
+	 * blocking Pwait() will follow: in order not to block forever, we must
+	 * convert it to nonblocking.
+	 */
+	if (P->awaiting_pending_stops && P->pending_stops == 0 && block)
+		block = 0;
+
+	/*
+	 * Never do a blocking wait for a process we already know is dead or
+	 * stopped.
+	 */
+	if (P->state != PS_RUN)
+		block = 0;
+
 	info.si_pid = 0;
 	do
 	{
@@ -783,11 +800,19 @@ Ptrace(struct ps_prochandle *P, int stopped)
 		    (P->state == PS_STOP))
 			return P->state;
 
+		if (P->state == PS_DEAD)
+			return P->state;
+
 		state = P->state;
 		P->ptrace_halted = TRUE;
 		kill(P->pid, SIGSTOP);
+
 		P->pending_stops++;
-		Pwait(P, 1);
+		P->awaiting_pending_stops++;
+		while (P->pending_stops && P->state == PS_RUN)
+			Pwait(P, 1);
+		P->awaiting_pending_stops--;
+
 		return state;
 	}
 
@@ -800,11 +825,16 @@ Ptrace(struct ps_prochandle *P, int stopped)
 
 	P->ptraced = TRUE;
 	P->ptrace_halted = TRUE;
-	P->pending_stops++;
 
-	if (Pwait(P, 1) == -1)
-		goto err;
-	else if (P->state != PS_TRACESTOP) {
+	P->pending_stops++;
+	P->awaiting_pending_stops++;
+	while (P->pending_stops && P->state == PS_RUN) {
+		if (Pwait(P, 1) == -1)
+			goto err;
+	}
+	P->awaiting_pending_stops--;
+
+	if (P->state != PS_TRACESTOP) {
 		err = ECHILD;
 		goto err2;
 	}
